old_traffic_prediction/requirements.txt
matplotlib
Pillow
pandas
apscheduler
bs4
numpy